Move models tests to their own module
